backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
azure-ai-documentintelligence==1.0.2
azure-core==1.33.0
azureopenai==0.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.0.14
django-cors-headers==4.7.0
frozenlist==1.5.0
idna==3.10
isodate==0.7.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mssql-django==1.5
multidict==6.3.2
propcache==0.3.1
pyodbc==5.2.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
yarl==1.19.0
